+++ src/test/static/namespaces/namespace-dynamic-named.js
@@ -5,6 +5,14 @@ DynamicNamespace['ArrayNotation'] = {
   foo: 'bar'
 };
 
+var baz = 'abc';
+/**
+ * @namespace DynamicNamespace.DynamicArrayNotation
+ */
+DynamicNamespace[baz] = {
+  foo: 'bar'
+};
+
 /**
  * @namespace DynamicNamespace.Aliased
  */