+++ src/test/static/namespaces/namespace-dynamic-unnamed.js
@@ -4,3 +4,11 @@
 DynamicNamespace['ArrayNotation'] = {
   foo: 'bar'
 };
+
+var baz = 'abc';
+/**
+ * @namespace
+ */
+DynamicNamespace[baz] = {
+  foo: 'bar'
+};